test(uport): tests for request and connect
moved integration tests into test/uportWeb3
